rename find to claim
