+++ src/test/kotlin/org/wfanet/measurement/kingdom/deploy/gcloud/spanner/writers/BUILD.bazel
@@ -130,8 +130,8 @@ spanner_emulator_test(
 )
 
 spanner_emulator_test(
-    name = "FindReadyExchangeStepTest",
-    srcs = ["FindReadyExchangeStepTest.kt"],
-    test_class = "org.wfanet.measurement.kingdom.deploy.gcloud.spanner.writers.FindReadyExchangeStepTest",
+    name = "ClaimReadyExchangeStepTest",
+    srcs = ["ClaimReadyExchangeStepTest.kt"],
+    test_class = "org.wfanet.measurement.kingdom.deploy.gcloud.spanner.writers.ClaimReadyExchangeStepTest",
     deps = _TEST_DEPS,
 )
